:hammer:refactor: reorganize auth module directory structure
- Restructure  by creating specific subdirectories based on functionality:
  -  for user and professional verification processes
  -  for user profile management
  -  for functionalities related to kit selection, management, and validation
- Simplify navigation and eliminate redundancy while keeping functionality clear
- Enhance modularization and preparation for future expansions
