--- app/src/main/java/com/example/autoclean/ui/auth/CleaningKitPaymentFragment.kt
@@ -1,23 +1,0 @@
-package com.example.autoclean.ui.auth
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.autoclean.R
-
-
-
-
-class CleaningKitPaymentFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_cleaning_kit_payment, container, false)
-    }
-
-}
