+++ app/src/main/java/com/example/autoclean/ui/auth/profile/ProfileFragment.kt
@@ -1,4 +1,4 @@
-package com.example.autoclean.ui.auth
+package com.example.autoclean.ui.auth.profile
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,19 +7,18 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
 import com.example.autoclean.R
-import com.example.autoclean.databinding.FragmentCodeVerificationBinding
-import com.example.autoclean.databinding.FragmentProfileProfissionalBinding
+import com.example.autoclean.databinding.FragmentProfileBinding
 
 
-class CodeVerificationFragment : Fragment() {
-    private var _binding: FragmentCodeVerificationBinding? = null
+class ProfileFragment : Fragment() {
+    private var _binding: FragmentProfileBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding=FragmentCodeVerificationBinding.inflate(inflater, container, false)
+       _binding = FragmentProfileBinding.inflate(inflater, container, false)
         return binding.root
     }
 
@@ -29,11 +28,15 @@ class CodeVerificationFragment : Fragment() {
         initListeners()
     }
 
-    private fun initListeners(){
-        binding.btnContinue.setOnClickListener{
-            findNavController().navigate(R.id.action_codeVerificationFragment_to_profileProfissionalFragment)
+    private fun initListeners() {
+
+        val navigateToRegister: (View) -> Unit = {
+            findNavController().navigate(R.id.action_profileFragment_to_registerFragment)
         }
+        binding.btnClient.setOnClickListener(navigateToRegister)
+        binding.btnProfissional.setOnClickListener(navigateToRegister)
     }
+
     override fun onDestroyView() {
         super.onDestroyView()
         _binding = null
