+++ app/src/main/java/com/example/autoclean/ui/auth/verification/ProfessionalVerificationFragment.kt
@@ -1,4 +1,4 @@
-package com.example.autoclean.ui.auth
+package com.example.autoclean.ui.auth.verification
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,17 +7,17 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
 import com.example.autoclean.R
-import com.example.autoclean.databinding.FragmentPhoneVerificationStartBinding
+import com.example.autoclean.databinding.FragmentProfessionalVerificationBinding
 
-class PhoneVerificationStartFragment : Fragment() {
-    private var _binding: FragmentPhoneVerificationStartBinding? = null
+class ProfessionalVerificationFragment : Fragment() {
+    private var _binding: FragmentProfessionalVerificationBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentPhoneVerificationStartBinding.inflate(inflater, container, false)
+        _binding = FragmentProfessionalVerificationBinding.inflate(inflater, container, false)
         return binding.root
     }
 
@@ -25,13 +25,18 @@ class PhoneVerificationStartFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         initListeners()
+        updateUserName("Carlos")
     }
 
-    private fun initListeners(){
+    private fun initListeners() {
         binding.btnContinue.setOnClickListener{
-            findNavController().navigate(R.id.action_phoneVerificationStartFragment_to_codeVerificationFragment)
-        }
+            findNavController().navigate(R.id.action_professionalVerificationFragment_to_cleaningkitmanagerFragment)
 
+        }
+    }
+
+    private fun updateUserName(userName: String) {
+        binding.nameProfile.text = "Olá, $userName"
     }
 
     override fun onDestroyView() {
