+++ app/src/main/res/layout/fragment_cleaning_kit.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.auth.CleaningKitFragment">
+    tools:context=".ui.auth.kits.CleaningKitFragment">
 
     <FrameLayout
         android:layout_width="match_parent"
