+++ app/src/main/res/layout/fragment_cleaningkitmanager.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    tools:context=".ui.auth.CleaningkitmanagerFragment">
+    tools:context=".ui.auth.kits.CleaningkitmanagerFragment">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/viewPager"
