+++ app/src/main/res/layout/fragment_code_verification.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.auth.PhoneVerificationStartFragment">
+    tools:context=".ui.auth.verification.PhoneVerificationStartFragment">
 
     <FrameLayout
         android:layout_width="match_parent"
