+++ app/src/main/res/layout/fragment_kit_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.auth.KitDetailsFragment">
+    tools:context=".ui.auth.kits.KitDetailsFragment">
 
     <FrameLayout
         android:layout_width="match_parent"
