+++ app/src/main/res/layout/fragment_phone_verification_start.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.auth.PhoneVerificationStartFragment">
+    tools:context=".ui.auth.verification.PhoneVerificationStartFragment">
 
     <FrameLayout
         android:layout_width="match_parent"
