+++ app/src/main/res/layout/fragment_professional_verification.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.auth.ProfessionalVerificationFragment">
+    tools:context=".ui.auth.verification.ProfessionalVerificationFragment">
 
     <FrameLayout
         android:layout_width="match_parent"
