+++ app/src/main/res/layout/fragment_profile_profissional.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.auth.ProfileProfissionalFragment">
+    tools:context=".ui.auth.profile.ProfileProfissionalFragment">
 
     <FrameLayout
         android:id="@+id/frameLayout"
