+++ app/src/main/res/layout/fragment_validation_kit.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.auth.ValidationKitFragment">
+    tools:context=".ui.auth.kits.ValidationKitFragment">
 
     <FrameLayout
         android:layout_width="match_parent"
